Initial public release (v0.0.3)
This is the initial public release of samctr, a cli program for
launching apptainer instances configured for building software and
modules for scientific software that is intended for distribution using
cvmfs. More specifically, it is intended for use with software build on
top of EESSI <https://eessi.io>. The program itself is based on the
eessi_container.sh script written by Thomas Roeblitz (@trz42).